init refactor base on discussion
