+++ nn_meter/builder/kernel_predictor_builder/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 from .data_sampler import Generator, generate_config_sample
-from .predictor_builder import build_predictor
+from .predictor_builder import build_predictor_by_data
