+++ nn_meter/builder/kernel_predictor_builder/data_sampler/prior_zoo/modelzoo_convs.csv
